Added constant columns and nonsingular variable set to DescriptiveStatsAction.
