+++ tetrad-gui/src/main/java/edu/cmu/tetradapp/editor/DataEditor.java
@@ -646,7 +646,7 @@ public final class DataEditor extends JPanel implements KnowledgeEditable,
         tools.add(new PlotMatrixAction(this));
         tools.add(new QQPlotAction(this));
         tools.add(new NormalityTestAction(this));
-        tools.add(new DescriptiveStatsAction2(this));
+        tools.add(new DescriptiveStatsAction(this));
 
         final int vkBackSpace = KeyEvent.VK_BACK_SPACE;
         final int vkDelete = KeyEvent.VK_DELETE;
